added entity menu to editor and support for entity class in rendering
In order to preserve data, I needed to store entity data in the editor, so i created a submenu for entity parameters in the editor which is saved as part of the grid. These params only decide which way the model is facing for now, but will be used for functionality later.

Recognizing the necessity for UI integration with entity interaction, the next step is to work on the player's interface and conceive a general theming and layout for the UI.