ironed out ui theming and components
created more components for the ui and decided on a theme. also added an interact prompt that pops up (no functionality yet) and a working minimap component.

next step is to make ui components draggable in-game, and then add a taskbar to the bottom of the ui that indicates whether components are open and if not allows to click on the icon to open them, and also add a close button to component windows