created input handling and vfx for attacks
attacks receive a down input and wait for an up input to perform an attack in that direction.

next step is to add the actual hit registration, enemy reactions, and a part system for enemy objects where enemy parts have their own hp and affect what actions an enemy is allowed to perform.