added breakable groups+parts
created data classes to assign multiple breakable parts on a cell and trigger a response when all are broken, works with items and walls without response implementation. also did some enemy bugfixing but map is still buggy.

plan for next week is to implement the rest of the basic UI menus (equipment, status, items, currency and some way to spend it) to give progression to the gameplay loop