equipment ui is functional
can now change equipment with the component and can place breakable walls on the editor too. ui as a whole is pretty slapshod and needs a revamp sometime in the future.

plan for tomorrow is to work a bit on combat feel and progression systems (currency drop from enemies, maybe start the skill tree component)