Added unlocking skills and stat display to skilltree
Can now purchase skills on the skilltree when requirements are met, though does not apply the effects yet. Stat changes are also previewed through a new stat window on the skill tree component. Also created a cursor overlay to show clickable things. Plan for tomorrow is to make skill effects actually apply to the player upon acquisition and create an implementation for active skills