Added controls and options for adding props to editor
Extra info in the editor to place a prop on any cell and give it an orientation when it is rendered in the scene. also made a shortcut to easily place preset props with one click. fixed a couple of rendering bugs too, but enemy display on the map is still buggy. Plan for tomorrow is to fix enemy icons on map, and add a warning indicator for when an enemy is charging an attack in an adjacent tile