imported model and set it up to be playable
created animations for the rigged bat model, set up enemy script and scriptableobject for its attack action, and adjusted handling a bit to work with the new imported model structure. also added a ragdall effect on death in lieu of making death animations for every enemy type. plan for tomorrow is to create the half-golem enemy and start improving enemy behaviors to be more intelligent.