created a new environment and area for the tutorial
created an office-themed environment for the tutorial area, with its own tileset and props. mostly finished but could use some polish. plan for tomorrow is to polish up the area and implement tutorial functionality (bool for starting in tutorial area, mechanics for requiring  new equipment to leave tutorial and render/go to the proper areas